refactor: Sign in filter refactoring
